cfm: Eight byte MPIDs in extended mode.
802.1ag only allows for MPIDs in the range [1, 8191].  This is
restrictive enough to make assignment of MPIDs to instances of OVS
awkward.  This patch allows eight byte MPIDs when running in
extended mode.

Bug #7014.

+++ lib/cfm.h
@@ -25,7 +25,7 @@ struct flow;
 struct ofpbuf;
 
 struct cfm_settings {
-    uint16_t mpid;              /* The MPID of this CFM. */
+    uint64_t mpid;              /* The MPID of this CFM. */
     int interval;               /* The requested transmission interval. */
     bool extended;              /* Run in extended mode. */
 };
